Extract shared rate-limit readout from session detail
